Rupert fixes, abstract parent refactor
