--- src/main/resources/assets/plushables/blockstates/rupert_box.json
@@ -1,20 +1,0 @@
-{
-  "variants": {
-    "facing=north": {
-      "y": 0,
-      "model": "plushables:block/rupert_box"
-    },
-    "facing=south": {
-      "y": 180,
-      "model": "plushables:block/rupert_box"
-    },
-    "facing=east": {
-      "y": 90,
-      "model": "plushables:block/rupert_box"
-    },
-    "facing=west": {
-      "y": 270,
-      "model": "plushables:block/rupert_box"
-    }
-  }
-}
