+++ src/main/resources/assets/plushables/lang/en_us.json
@@ -15,5 +15,5 @@
   "block.plushables.unicorn_plushable": "Plushable §c§lU§6§ln§e§li§a§lc§b§lo§9§lr§5§ln",
   "block.plushables.dragon_plushable": "Plushable §4§lDragon",
   "block.plushables.raptor_plushable": "Plushable §9§lVelociraptor",
-  "block.plushables.rupert_box": "Boxed Plushable §8§lRupert"
+  "block.plushables.rupert_plushable": "Plushable §8§lRupert"
 }
