--- src/main/resources/assets/plushables/models/block/rupert_box.json
@@ -1,190 +1,0 @@
-{
-	"credit": "Made with Blockbench",
-	"textures": {
-      "0": "plushables:block/rupert_box_texture",
-      "particle": "plushables:block/blank"
-	},
-	"elements": [
-		{
-			"from": [2, 2, 2],
-			"to": [4, 12, 14],
-			"faces": {
-				"north": {"uv": [15.5, 0, 16, 2.5], "texture": "#0"},
-				"east": {"uv": [13, 0, 16, 2.5], "texture": "#0"},
-				"south": {"uv": [15.5, 0, 16, 2.5], "texture": "#0"},
-				"west": {"uv": [13, 0, 16, 2.5], "texture": "#0"},
-				"up": {"uv": [15.5, 0, 16, 2.5], "texture": "#0"},
-				"down": {"uv": [15.5, 0, 16, 2.5], "texture": "#0"}
-			}
-		},
-		{
-			"from": [12, 2, 12],
-			"to": [14, 12, 14],
-			"faces": {
-				"north": {"uv": [14.5, 0.5, 15, 3], "texture": "#0"},
-				"east": {"uv": [15, 0.5, 14.5, 3], "texture": "#0"},
-				"south": {"uv": [14.5, 0.5, 15, 3], "texture": "#0"},
-				"west": {"uv": [14.5, 0.5, 15, 3], "texture": "#0"},
-				"up": {"uv": [14.5, 0.5, 15, 3], "texture": "#0"},
-				"down": {"uv": [14.5, 0.5, 15, 3], "texture": "#0"}
-			}
-		},
-		{
-			"from": [13, 2, 2],
-			"to": [14, 12, 3],
-			"faces": {
-				"north": {"uv": [14.5, 0.5, 15, 3], "texture": "#0"},
-				"east": {"uv": [15, 0.5, 14.5, 3], "texture": "#0"},
-				"south": {"uv": [14.5, 0.5, 15, 3], "texture": "#0"},
-				"west": {"uv": [14.5, 0.5, 15, 3], "texture": "#0"},
-				"up": {"uv": [14.5, 0.5, 15, 3], "texture": "#0"},
-				"down": {"uv": [14.5, 0.5, 15, 3], "texture": "#0"}
-			}
-		},
-		{
-			"from": [4, 2, 12],
-			"to": [12, 12, 14],
-			"faces": {
-				"north": {"uv": [14, 0, 16, 2.5], "texture": "#0"},
-				"east": {"uv": [15.5, 0, 16, 2.5], "texture": "#0"},
-				"south": {"uv": [0, 11, 4, 16], "texture": "#0"},
-				"west": {"uv": [15.5, 0, 16, 2.5], "texture": "#0"},
-				"up": {"uv": [14, 0, 16, 0.5], "texture": "#0"},
-				"down": {"uv": [14, 0, 16, 0.5], "texture": "#0"}
-			}
-		},
-		{
-			"from": [2, 0, 2],
-			"to": [14, 2, 14],
-			"faces": {
-				"north": {"uv": [0, 0, 4, 1], "texture": "#0"},
-				"east": {"uv": [0, 0, 4, 1], "texture": "#0"},
-				"south": {"uv": [0, 0, 4, 1], "texture": "#0"},
-				"west": {"uv": [0, 0, 4, 1], "texture": "#0"},
-				"up": {"uv": [0, 0, 1, 1], "texture": "#0"},
-				"down": {"uv": [0, 0, 1, 1], "texture": "#0"}
-			}
-		},
-		{
-			"from": [2, 12, 2],
-			"to": [14, 14, 14],
-			"faces": {
-				"north": {"uv": [0, 0, 4, 1], "texture": "#0"},
-				"east": {"uv": [0, 0, 4, 1], "texture": "#0"},
-				"south": {"uv": [0, 0, 4, 1], "texture": "#0"},
-				"west": {"uv": [0, 0, 4, 1], "texture": "#0"},
-				"up": {"uv": [8, 0, 10, 2], "texture": "#0"},
-				"down": {"uv": [0, 0, 1, 1], "texture": "#0"}
-			}
-		},
-		{
-			"from": [10, 9, 2],
-			"to": [13, 12, 4],
-			"faces": {
-				"north": {"uv": [6, 3, 8.5, 5.5], "texture": "#0"},
-				"east": {"uv": [6, 14, 6.5, 14.5], "texture": "#0"},
-				"south": {"uv": [4.5, 0, 5.5, 1], "texture": "#0"},
-				"west": {"uv": [6, 14, 6.5, 14.5], "texture": "#0"},
-				"up": {"uv": [6, 14, 7, 14.5], "texture": "#0"},
-				"down": {"uv": [6, 14, 7, 14.5], "texture": "#0"}
-			}
-		},
-		{
-			"from": [4, 2, 2],
-			"to": [7, 5, 4],
-			"faces": {
-				"north": {"uv": [10, 7, 12.5, 9.5], "texture": "#0"},
-				"east": {"uv": [6, 14, 6.5, 14.5], "texture": "#0"},
-				"south": {"uv": [4.5, 0, 5.5, 1], "texture": "#0"},
-				"west": {"uv": [6, 14, 6.5, 14.5], "texture": "#0"},
-				"up": {"uv": [6, 14, 7, 14.5], "texture": "#0"},
-				"down": {"uv": [6, 14, 7, 14.5], "texture": "#0"}
-			}
-		},
-		{
-			"from": [12, 8, 8],
-			"to": [14, 12, 12],
-			"faces": {
-				"north": {"uv": [6, 11.5, 6.5, 15.5], "texture": "#0"},
-				"east": {"uv": [10, 9.5, 12.5, 12], "texture": "#0"},
-				"south": {"uv": [6, 11.5, 6.5, 15.5], "texture": "#0"},
-				"west": {"uv": [4.5, 0, 5.5, 1], "texture": "#0"},
-				"up": {"uv": [6, 11.5, 6.5, 15.5], "texture": "#0"},
-				"down": {"uv": [6, 11.5, 6.5, 15.5], "texture": "#0"}
-			}
-		},
-		{
-			"from": [12, 2, 3],
-			"to": [14, 6, 7],
-			"faces": {
-				"north": {"uv": [6, 11.5, 6.5, 15.5], "texture": "#0"},
-				"east": {"uv": [13, 13.5, 15.5, 16], "texture": "#0"},
-				"south": {"uv": [6, 11.5, 6.5, 15.5], "texture": "#0"},
-				"west": {"uv": [4.5, 0, 5.5, 1], "texture": "#0"},
-				"up": {"uv": [6, 11.5, 6.5, 15.5], "texture": "#0"},
-				"down": {"uv": [6, 11.5, 6.5, 15.5], "texture": "#0"}
-			}
-		},
-		{
-			"from": [4, 1.5, 2.75],
-			"to": [13, 10.5, 11.75],
-			"faces": {
-				"north": {"uv": [0, 5, 5, 10], "texture": "#0"},
-				"east": {"uv": [0, 5, 5, 10], "texture": "#0"},
-				"south": {"uv": [0, 5, 5, 10], "texture": "#0"},
-				"west": {"uv": [0, 5, 5, 10], "texture": "#0"},
-				"up": {"uv": [0, 1.5, 5, 3.5], "texture": "#0"},
-				"down": {"uv": [0, 1.5, 5, 3.5], "texture": "#0"}
-			}
-		}
-	],
-	"display": {
-		"thirdperson_righthand": {
-			"rotation": [67, 42, 0],
-			"translation": [0.5, 1.25, 1.25],
-			"scale": [0.3, 0.3, 0.3]
-		},
-		"thirdperson_lefthand": {
-			"rotation": [67, 42, 0],
-			"translation": [0.5, 1.25, 1.25],
-			"scale": [0.3, 0.3, 0.3]
-		},
-		"firstperson_righthand": {
-			"rotation": [-1, -180, 0],
-			"translation": [-9, 1, -10]
-		},
-		"firstperson_lefthand": {
-			"rotation": [0, 118, 0],
-			"translation": [7.5, 1, -9.75]
-		},
-		"ground": {
-			"translation": [0, 1.5, 0],
-			"scale": [0.6, 0.6, 0.6]
-		},
-		"gui": {
-			"rotation": [21, -132, 0],
-			"translation": [0, 0.75, 0],
-			"scale": [0.8, 0.8, 0.8]
-		},
-		"head": {
-			"translation": [0, 14, 0]
-		},
-		"fixed": {
-			"translation": [0, 0, -4.25]
-		}
-	},
-	"groups": [
-		0,
-		1,
-		2,
-		3,
-		4,
-		5,
-		{
-			"name": "glass",
-			"origin": [0, 0, 0],
-			"color": 0,
-			"children": [6, 7, 8, 9, 10]
-		}
-	]
-}
