--- src/main/resources/assets/plushables/models/item/rupert_box.json
@@ -1,4 +1,0 @@
-{
-  "parent": "plushables:block/rupert_box",
-  "gui_light": "front"
-}
