+++ src/main/resources/assets/plushables/models/item/rupert_plushable.json
