+++ src/main/resources/data/plushables/recipes/rupert_plushable.json
@@ -12,7 +12,7 @@
     }
   ],
   "output": {
-    "item": "plushables:rupert_box",
+    "item": "plushables:rupert_plushable",
     "count": 1
   }
 }
